+++ vvf_mobile/index.js
@@ -6,6 +6,6 @@ import {AppRegistry} from 'react-native';
 import App from './App';
 import './src/libs/navigations/gesture-handler';
 import {name as appName} from './app.json';
-import './i18config'
+import './i18config';
 
 AppRegistry.registerComponent(appName, () => App);
