+++ vvf_mobile/src/services/postService.ts
@@ -1,5 +1,5 @@
 import ResponseDTO from '@data/responseDTO';
-import axiosInstance from './apis/axios';
+import axiosInstance from '../libs/apis/axios';
 import {Post} from '@data/post';
 import {Alert} from 'react-native';
