+++ vvf_mobile/src/ui/main/event/EventDetailScreen.tsx
@@ -10,7 +10,7 @@ import {
   Alert,
 } from 'react-native';
 import {RouteProp, useRoute} from '@react-navigation/native';
-import axiosInstance from 'src/services/apis/axios';
+import axiosInstance from 'src/libs/apis/axios';
 import ResponseDTO from '@data/responseDTO';
 import ImageInfo from './components/ImageInfo';
 import {Event} from '@data/event';
@@ -20,6 +20,7 @@ import {dateToString} from 'src/utils/dateTimeUtil';
 import EventScheduleItem from './components/EventScheduleItem';
 import Divider from '@components/Divider';
 import {useTranslation} from 'react-i18next';
+import FloattingContainer from './components/FloattingContainer';
 
 type EventDetailScreenParams = {
   eventId: string;
@@ -42,6 +43,7 @@ const EventDetailScreen = () => {
           `/events/get?eventId=${requestEventId}`,
         );
         if (res.data?.data) {
+          console.log(res.data.data);
           setEventInfo(res.data.data);
         } else {
           setError('Event data is unavailable.');
@@ -87,52 +89,57 @@ const EventDetailScreen = () => {
   }
 
   return (
-    <ScrollView contentContainerStyle={styles.scrollContainer}>
-      <View>
-        {eventInfo.imgUrl && (
-          <ImageInfo imageUrl={eventInfo.imgUrl} style={styles.imageInfo} />
+    <View style={styles.wrapper}>
+      <ScrollView contentContainerStyle={styles.scrollContainer}>
+        <View>
+          {eventInfo.imgUrl && (
+            <ImageInfo imageUrl={eventInfo.imgUrl} style={styles.imageInfo} />
+          )}
+
+          <View style={styles.eventTitleInfoContainer}>
+            <Text style={[styles.eventTitleInfoText]}>
+              {`${dateToString(eventInfo.startDate, 'DD/MM/YYYY hh:mm')} | ${
+                eventInfo.location
+              }`}
+            </Text>
+            <Text style={[styles.eventTitleInfoText, styles.title]}>
+              {eventInfo.title}
+            </Text>
+            <FilledButtonComponent
+              style={styles.reserveBtn}
+              onPress={handleReserve}
+              title={t('reserve')}
+              textStyle={styles.reserveBtnLabel}
+            />
+          </View>
+        </View>
+        <SectionTitle title={t('time_and_location')} />
+        <View style={styles.timeAndLocationContainer}>
+          <Text>Date: {dateToString(eventInfo.startDate, 'DD/MM/YYYY')}</Text>
+          <Text>Time: {eventInfo.startTime}</Text>
+          <Text>Location: {eventInfo.location}</Text>
+        </View>
+
+        <SectionTitle title={t('about_the_event')} />
+        {eventInfo.description && (
+          <HtmlComponent html={eventInfo.description} />
         )}
 
-        <View style={styles.eventTitleInfoContainer}>
-          <Text style={[styles.eventTitleInfoText]}>
-            {`${dateToString(eventInfo.startDate, 'DD/MM/YYYY hh:mm')} | ${
-              eventInfo.location
-            }`}
-          </Text>
-          <Text style={[styles.eventTitleInfoText, styles.title]}>
-            {eventInfo.title}
-          </Text>
-          <FilledButtonComponent
-            style={styles.reserveBtn}
-            onPress={handleReserve}
-            title={t('reserve')}
-            textStyle={styles.reserveBtnLabel}
-          />
+        <SectionTitle title={t('schedule')} />
+        <Text style={styles.subTitle}>
+          {t('may_change_according_to_instructor')}
+        </Text>
+        <View style={styles.eventScheduleList}>
+          {eventInfo.schedules?.map((schedule, index) => (
+            <View key={index}>
+              <EventScheduleItem item={schedule} />
+              <Divider type="horizontal" />
+            </View>
+          ))}
         </View>
-      </View>
-      <SectionTitle title={t('time_and_location')} />
-      <View style={styles.timeAndLocationContainer}>
-        <Text>Date: {dateToString(eventInfo.startDate, 'DD/MM/YYYY')}</Text>
-        <Text>Time: {eventInfo.startTime}</Text>
-        <Text>Location: {eventInfo.location}</Text>
-      </View>
-
-      <SectionTitle title={t('about_the_event')} />
-      {eventInfo.description && <HtmlComponent html={eventInfo.description} />}
-
-      <SectionTitle title={t('schedule')} />
-      <Text style={styles.subTitle}>
-        {t('may_change_according_to_instructor')}
-      </Text>
-      <View style={styles.eventScheduleList}>
-        {eventInfo.schedules?.map((schedule, index) => (
-          <View key={index}>
-            <EventScheduleItem item={schedule} />
-            <Divider type="horizontal" />
-          </View>
-        ))}
-      </View>
-    </ScrollView>
+      </ScrollView>
+      <FloattingContainer />
+    </View>
   );
 };
 
@@ -141,6 +148,10 @@ const SectionTitle = ({title}: {title: string}) => (
 );
 
 const styles = StyleSheet.create({
+  wrapper: {
+    flex: 1,
+    position: 'relative',
+  },
   scrollContainer: {
     flexGrow: 1,
     paddingHorizontal: 10,
